[add]  add asset UI

+++ Assets/EnemiesHealth.cs
@@ -5,14 +5,14 @@ using UnityEngine;
 public class EnemiesHealth : MonoBehaviour
 {
     public Animator animator;
-    [SerializeField] public int maxHealth = 100;
-    private int currentHealth;
+    [SerializeField] public float maxHealth = 100;
+    private float currentHealth;
    private void Start()
     {
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
